6422 | 7 files | M DumpSource2/schemas/client/CDOTA_Modifier_BlackDragon_Fireball_Thinker.h,  M DumpSource2/schemas/client/C_DOTA_Ability_Meepo_Megameepo_Fling.h,  M DumpSource2/schemas/server/CDOTA_Ability_Meepo_Megameepo_Fling.h,  M DumpSource2/schemas/server/CDOTA_Modifier_BlackDragon_Fireball_Thinker.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/18588119/

+++ DumpSource2/schemas/client/CDOTA_Modifier_BlackDragon_Fireball_Thinker.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_BlackDragon_Fireball_Thinker : public CDOTA_Buff
 {
 	float32 burn_interval;
-	int32 radius;
-	int32 damage;
+	float32 radius;
+	float32 damage;
 };
